- KOGITO-2992: Context menu actions       - Selected node(s) is(are) added to the new DRD       - Fix overlapping breadcrumbs (issue reported by Jan)

+++ kie-wb-common-dmn/kie-wb-common-dmn-client/src/main/java/org/kie/workbench/common/dmn/client/editors/drd/DRDContextMenu.java
@@ -18,13 +18,16 @@ package org.kie.workbench.common.dmn.client.editors.drd;
 
 import java.util.Collection;
 import java.util.List;
+import java.util.function.Predicate;
 
 import javax.enterprise.context.ApplicationScoped;
 import javax.inject.Inject;
 
 import elemental2.dom.DomGlobal;
 import elemental2.dom.HTMLElement;
+import org.kie.workbench.common.dmn.api.property.dmn.Id;
 import org.kie.workbench.common.dmn.client.docks.navigator.drds.DMNDiagramTuple;
+import org.kie.workbench.common.dmn.client.docks.navigator.drds.DMNDiagramsSession;
 import org.kie.workbench.common.dmn.client.editors.contextmenu.ContextMenu;
 import org.kie.workbench.common.stunner.core.client.i18n.ClientTranslationService;
 import org.kie.workbench.common.stunner.core.graph.Edge;
@@ -43,14 +46,17 @@ public class DRDContextMenu {
     private final ClientTranslationService translationService;
     private final ContextMenu contextMenu;
     private final DRDContextMenuService drdContextMenuService;
+    private final DMNDiagramsSession dmnDiagramsSession;
 
     @Inject
     public DRDContextMenu(final ContextMenu contextMenu,
                           final ClientTranslationService translationService,
-                          final DRDContextMenuService drdContextMenuService) {
+                          final DRDContextMenuService drdContextMenuService,
+                          final DMNDiagramsSession dmnDiagramsSession) {
         this.contextMenu = contextMenu;
         this.translationService = translationService;
         this.drdContextMenuService = drdContextMenuService;
+        this.dmnDiagramsSession = dmnDiagramsSession;
     }
 
     public String getTitle() {
@@ -76,15 +82,38 @@ public class DRDContextMenu {
                                     true,
                                     () -> drdContextMenuService.addToNewDRD(selectedNodes));
 
-        getDiagrams().forEach(dmnDiagram -> {
-            contextMenu.addTextMenuItem(translationService.getValue(DRDACTIONS_CONTEXT_MENU_ACTIONS_ADD_TO) + " " + getDiagramName(dmnDiagram),
-                                        true,
-                                        () -> drdContextMenuService.addToExistingDRD(dmnDiagram, selectedNodes));
-        });
+        getDiagrams()
+                .stream()
+                .filter(excludeGlobalGraphPredicate())
+                .filter(excludeCurrentDRDPredicate())
+                .forEach(dmnDiagram ->
+                                 contextMenu.addTextMenuItem(translationService.getValue(DRDACTIONS_CONTEXT_MENU_ACTIONS_ADD_TO) + " " + getDiagramName(dmnDiagram),
+                                                             true,
+                                                             () -> drdContextMenuService.addToExistingDRD(dmnDiagram, selectedNodes))
+                );
 
-        contextMenu.addTextMenuItem(translationService.getValue(DRDACTIONS_CONTEXT_MENU_ACTIONS_REMOVE),
-                                    true,
-                                    () -> drdContextMenuService.removeFromCurrentDRD(selectedNodes));
+        if (!dmnDiagramsSession.isGlobalGraphSelected()) {
+            contextMenu.addTextMenuItem(translationService.getValue(DRDACTIONS_CONTEXT_MENU_ACTIONS_REMOVE),
+                                        true,
+                                        () -> drdContextMenuService.removeFromCurrentDRD(selectedNodes));
+        }
+    }
+
+    private Predicate<DMNDiagramTuple> excludeCurrentDRDPredicate() {
+        return dmnDiagramTuple -> {
+            final Id dmnDiagramId = dmnDiagramTuple.getDMNDiagram().getId();
+            return dmnDiagramsSession
+                    .getCurrentDMNDiagramElement()
+                    .map(dmnDiagramElement -> !dmnDiagramId.equals(dmnDiagramElement.getId()))
+                    .orElse(false);
+        };
+    }
+
+    private Predicate<DMNDiagramTuple> excludeGlobalGraphPredicate() {
+        return dmnDiagramTuple -> {
+            final Id dmnDiagramId = dmnDiagramTuple.getDMNDiagram().getId();
+            return !dmnDiagramId.equals(dmnDiagramsSession.getDRGDiagramElement().getId());
+        };
     }
 
     private String getDiagramName(final DMNDiagramTuple dmnDiagram) {
